processing-ec2/requirements.txt
fastapi==0.117.1
pefile==2024.8.26
python_magic==0.4.27
